+++ masterstudy-lms-learning-management-system/includes/Http/Controllers/Media/DeleteController.php
@@ -14,6 +14,16 @@ class DeleteController {
 			);
 		}
 
+		if ( ! current_user_can( 'delete_post', $media_id ) ) {
+			return new \WP_REST_Response(
+				array(
+					'error_code' => 'delete_media_access_error',
+					'message'    => esc_html__( 'You do not have permission to delete this template', 'masterstudy-lms-learning-management-system' ),
+				),
+				403
+			);
+		}
+
 		return new \WP_REST_Response(
 			array(
 				'status' => (bool) wp_delete_attachment( $media_id ),
